Fix: Use full images for the gallery instead of thumbnails [ED-12088]
