+++ modules/ai/assets/js/editor/components/prompt-history/helpers/image-helpers.js
@@ -1,0 +1,4 @@
+export const getImageThumbnailURL = ( imageURL ) => {
+	// FIXME: Use URL from meta once available
+	return `https://my.stg.elementor.red/ai/thumb/?o=${ btoa( imageURL, false ) }`;
+};
